+++ src/firefly_synth/firefly_synth/matrix_cv.cpp
@@ -21,7 +21,8 @@ static int constexpr global_route_count = 20;
 static int constexpr max_route_count = std::max(voice_route_count, global_route_count);
 
 enum { section_main };
-enum { param_type, param_source, param_target, param_min, param_max };
+enum { transform_source_scratch, scratch_count };
+enum { param_type, param_source, param_target, param_offset, param_scale, param_min, param_max };
 enum { type_off, type_mul_abs, type_mul_rel, type_mul_stk, type_add_abs, type_add_rel, type_add_stk, type_ab_abs, type_ab_rel, type_ab_stk };
 
 extern void
@@ -275,7 +276,7 @@ cv_matrix_topo(
 
   int route_count = global ? global_route_count : voice_route_count;
   module_topo result(make_module(info,
-    make_module_dsp(stage, module_output::cv, 0, {
+    make_module_dsp(stage, module_output::cv, scratch_count, {
       make_module_dsp_output(false, make_topo_info("{3AEE42C9-691E-484F-B913-55EB05CFBB02}", "Output", 0, route_count)) }),
     make_module_gui(section, colors, pos, { 1, 1 })));
   
@@ -301,7 +302,7 @@ cv_matrix_topo(
 
   auto& main = result.sections.emplace_back(make_param_section(section_main,
     make_topo_tag("{A19E18F8-115B-4EAB-A3C7-43381424E7AB}", "Main"), 
-    make_param_section_gui({ 0, 0 }, { { 1 }, { gui_dimension::auto_size, 3, 4, -30, -30 } })));
+    make_param_section_gui({ 0, 0 }, { { 1 }, { gui_dimension::auto_size, 3, 4, -24, -24, -24, -24 } })));
   main.gui.scroll_mode = gui_scroll_mode::vertical;
   
   auto& type = result.params.emplace_back(make_param(
@@ -333,7 +334,6 @@ cv_matrix_topo(
   source.gui.submenu = source_matrix.submenu;
   source.info.description = std::string("All global CV and MIDI sources, plus for per-voice CV all per-voice CV sources, ") + 
     "MIDI note and velocity, and On-Note all global CV sources.";
-
   auto& target = result.params.emplace_back(make_param(
     make_topo_info("{94A037CE-F410-4463-8679-5660AFD1582E}", "Target", "Target", true, true, param_target, route_count),
     make_param_dsp_input(!global, param_automate::automate), make_domain_item(target_matrix.items, ""),
@@ -344,18 +344,33 @@ cv_matrix_topo(
   target.gui.item_enabled.auto_bind = true;
   target.info.description = "Any modulatable parameter of any audio module, audio matrix or (in case of per-voice) voice-in parameter.";
 
-  auto& min = result.params.emplace_back(make_param(
-    make_topo_info("{71E6F836-1950-4C8D-B62B-FAAD20B1FDBD}", "Min", "Min", true, true, param_min, route_count),
-    make_param_dsp_accurate(param_automate::modulate), make_domain_percentage_identity(0, 0, true),
+  auto& offset = result.params.emplace_back(make_param(
+    make_topo_info("{86ECE946-D554-4445-B8ED-2A7380C910E4}", "Offset", "Of", true, true, param_offset, route_count),
+    make_param_dsp_accurate(param_automate::modulate), make_domain_linear(-1, 1, 0, 2, ""),
     make_param_gui(section_main, gui_edit_type::knob, param_layout::vertical, { 0, 3 }, make_label_none())));
+  offset.gui.tabular = true;
+  offset.gui.bindings.enabled.bind_params({ param_type }, [](auto const& vs) { return vs[0] != type_off; });
+  offset.info.description = std::string("Source signal offset. Used to transform source before modulation is applied. ") +
+    "Useful to stretch things like midi note/velocity into the full [0, 1] range.";
+  auto& scale = result.params.emplace_back(make_param(
+    make_topo_info("{6564CE04-0AB8-4CDD-8F3D-E477DD1F4715}", "Scale", "Sc", true, true, param_scale, route_count),
+    make_param_dsp_accurate(param_automate::modulate), make_domain_linear(1, 32, 1, 2, ""),
+    make_param_gui(section_main, gui_edit_type::knob, param_layout::vertical, { 0, 4 }, make_label_none())));
+  scale.gui.tabular = true;
+  scale.gui.bindings.enabled.bind_params({ param_type }, [](auto const& vs) { return vs[0] != type_off; });
+  scale.info.description = std::string("Source signal multiplier. Used to transform source before modulation is applied. ") +
+    "Useful to stretch things like midi note/velocity into the full [0, 1] range.";
+  auto& min = result.params.emplace_back(make_param(
+    make_topo_info("{71E6F836-1950-4C8D-B62B-FAAD20B1FDBD}", "Min", "Mn", true, true, param_min, route_count),
+    make_param_dsp_accurate(param_automate::modulate), make_domain_percentage_identity(0, 0, true),
+    make_param_gui(section_main, gui_edit_type::knob, param_layout::vertical, { 0, 5 }, make_label_none())));
   min.gui.tabular = true;
   min.gui.bindings.enabled.bind_params({ param_type }, [](auto const& vs) { return vs[0] != type_off; });
   min.info.description = "Defines the bounds of the modulation effect. When min > max, modulation will invert.";
-
   auto& max = result.params.emplace_back(make_param(
-    make_topo_info("{DB3A5D43-95CB-48DC-97FA-984F55B57F7B}", "Max", "Max", true, true, param_max, route_count),
+    make_topo_info("{DB3A5D43-95CB-48DC-97FA-984F55B57F7B}", "Max", "Mx", true, true, param_max, route_count),
     make_param_dsp_accurate(param_automate::modulate), make_domain_percentage_identity(1, 0, true),
-    make_param_gui(section_main, gui_edit_type::knob, param_layout::vertical, { 0, 4 }, make_label_none())));
+    make_param_gui(section_main, gui_edit_type::knob, param_layout::vertical, { 0, 6 }, make_label_none())));
   max.gui.tabular = true;
   max.gui.bindings.enabled.bind_params({ param_type }, [](auto const& vs) { return vs[0] != type_off; });
   max.info.description = "Defines the bounds of the modulation effect. When min > max, modulation will invert.";
@@ -440,46 +455,59 @@ cv_matrix_engine::process(plugin_block& block)
     for(int f = block.start_frame; f < block.end_frame; f++)
       check_unipolar(source_curve[f]);
 
+    // pre-transform source signal
+    int this_module = _global? module_gcv_matrix: module_vcv_matrix;
+    auto const& scale_curve = block.state.own_accurate_automation[param_scale][r];
+    auto const& offset_curve = block.state.own_accurate_automation[param_offset][r];
+    auto& transformed_source = block.state.own_scratch[transform_source_scratch];
+    for (int f = block.start_frame; f < block.end_frame; f++)
+    {
+      float scale = block.normalized_to_raw_fast<domain_type::linear>(this_module, param_scale, scale_curve[f]);
+      float offset = block.normalized_to_raw_fast<domain_type::linear>(this_module, param_offset, offset_curve[f]);
+      transformed_source[f] = std::clamp((offset + source_curve[f]) * scale, 0.0f, 1.0f);
+    }
+
     // apply modulation
     auto const& min_curve = block.state.own_accurate_automation[param_min][r];
     auto const& max_curve = block.state.own_accurate_automation[param_max][r];
+    
     switch (type)
     {
     case type_mul_abs:
       for (int f = block.start_frame; f < block.end_frame; f++)
-        modulated_curve[f] *= min_curve[f] + (max_curve[f] - min_curve[f]) * source_curve[f];
+        modulated_curve[f] *= min_curve[f] + (max_curve[f] - min_curve[f]) * transformed_source[f];
       break;
     case type_mul_rel:
       for (int f = block.start_frame; f < block.end_frame; f++)
-        modulated_curve[f] = target_curve[f] + (min_curve[f] + (max_curve[f] - min_curve[f]) * source_curve[f]) * (1 - target_curve[f]);
+        modulated_curve[f] = target_curve[f] + (min_curve[f] + (max_curve[f] - min_curve[f]) * transformed_source[f]) * (1 - target_curve[f]);
       break;
     case type_mul_stk:
       for (int f = block.start_frame; f < block.end_frame; f++)
-        modulated_curve[f] = modulated_curve[f] + (min_curve[f] + (max_curve[f] - min_curve[f]) * source_curve[f]) * (1 - modulated_curve[f]);
+        modulated_curve[f] = modulated_curve[f] + (min_curve[f] + (max_curve[f] - min_curve[f]) * transformed_source[f]) * (1 - modulated_curve[f]);
       break;
     case type_add_abs:
       for (int f = block.start_frame; f < block.end_frame; f++)
-        modulated_curve[f] += min_curve[f] + (max_curve[f] - min_curve[f]) * source_curve[f];
+        modulated_curve[f] += min_curve[f] + (max_curve[f] - min_curve[f]) * transformed_source[f];
       break;
     case type_add_rel:
       for (int f = block.start_frame; f < block.end_frame; f++)
-        modulated_curve[f] += (1 - target_curve[f]) * (min_curve[f] + (max_curve[f] - min_curve[f]) * source_curve[f]);
+        modulated_curve[f] += (1 - target_curve[f]) * (min_curve[f] + (max_curve[f] - min_curve[f]) * transformed_source[f]);
       break;
     case type_add_stk:
       for (int f = block.start_frame; f < block.end_frame; f++)
-        modulated_curve[f] += (1 - modulated_curve[f]) * (min_curve[f] + (max_curve[f] - min_curve[f]) * source_curve[f]);
+        modulated_curve[f] += (1 - modulated_curve[f]) * (min_curve[f] + (max_curve[f] - min_curve[f]) * transformed_source[f]);
       break;
     case type_ab_abs:
       for (int f = block.start_frame; f < block.end_frame; f++)
-        modulated_curve[f] += unipolar_to_bipolar(min_curve[f] + (max_curve[f] - min_curve[f]) * source_curve[f]) * 0.5f;
+        modulated_curve[f] += unipolar_to_bipolar(min_curve[f] + (max_curve[f] - min_curve[f]) * transformed_source[f]) * 0.5f;
       break;
     case type_ab_rel:
       for (int f = block.start_frame; f < block.end_frame; f++)
-        modulated_curve[f] += (1 - std::fabs(0.5f - target_curve[f]) * 2.0f) * unipolar_to_bipolar(min_curve[f] + (max_curve[f] - min_curve[f]) * source_curve[f]) * 0.5f;
+        modulated_curve[f] += (1 - std::fabs(0.5f - target_curve[f]) * 2.0f) * unipolar_to_bipolar(min_curve[f] + (max_curve[f] - min_curve[f]) * transformed_source[f]) * 0.5f;
       break;
     case type_ab_stk:
       for (int f = block.start_frame; f < block.end_frame; f++)
-        modulated_curve[f] += (1 - std::fabs(0.5f - modulated_curve[f]) * 2.0f) * unipolar_to_bipolar(min_curve[f] + (max_curve[f] - min_curve[f]) * source_curve[f]) * 0.5f;
+        modulated_curve[f] += (1 - std::fabs(0.5f - modulated_curve[f]) * 2.0f) * unipolar_to_bipolar(min_curve[f] + (max_curve[f] - min_curve[f]) * transformed_source[f]) * 0.5f;
       break;
     default:
       assert(false);